Undo most of https://github.com/google/pybind11k/pull/30008 (does NOT build)
