+++ CMakeLists.txt
@@ -147,10 +147,8 @@ endif()
 
 # NB: when adding a header don't forget to also add it to setup.py
 set(PYBIND11_HEADERS
-    include/pybind11/detail/abi_platform_id.h
     include/pybind11/detail/class.h
     include/pybind11/detail/common.h
-    include/pybind11/detail/cross_extension_shared_state.h
     include/pybind11/detail/descr.h
     include/pybind11/detail/dynamic_raw_ptr_cast_if_possible.h
     include/pybind11/detail/function_record_pyobject.h
@@ -162,7 +160,6 @@ set(PYBIND11_HEADERS
     include/pybind11/detail/smart_holder_type_casters.h
     include/pybind11/detail/type_caster_base.h
     include/pybind11/detail/type_caster_odr_guard.h
-    include/pybind11/detail/type_map.h
     include/pybind11/detail/typeid.h
     include/pybind11/detail/value_and_holder.h
     include/pybind11/attr.h
